LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/fetch/private-network-access/iframe.tentative.https.window-expected.txt


PASS local to local: no preflight required.
PASS local to private: no preflight required.
PASS local to public: no preflight required.
FAIL private to local: failed preflight. assert_equals: expected "timeout" but got "loaded"
PASS private to local: missing CORS headers.
PASS private to local: missing PNA header.
FAIL private to local: success. assert_equals: expected "loaded" but got "timeout"
PASS private to private: no preflight required.
PASS private to public: no preflight required.
FAIL public to local: failed preflight. assert_equals: expected "timeout" but got "loaded"
PASS public to local: missing CORS headers.
PASS public to local: missing PNA header.
FAIL public to local: success. assert_equals: expected "loaded" but got "timeout"
FAIL public to private: failed preflight. assert_equals: expected "timeout" but got "loaded"
PASS public to private: missing CORS headers.
PASS public to private: missing PNA header.
FAIL public to private: success. assert_equals: expected "loaded" but got "timeout"
PASS public to public: no preflight required.
FAIL treat-as-public-address to local: failed preflight. assert_equals: expected "timeout" but got "loaded"
PASS treat-as-public-address to local: missing CORS headers.
PASS treat-as-public-address to local: missing PNA header.
FAIL treat-as-public-address to local: success. assert_equals: expected "loaded" but got "timeout"
FAIL treat-as-public-address to private: failed preflight. assert_equals: expected "timeout" but got "loaded"
PASS treat-as-public-address to private: missing CORS headers.
PASS treat-as-public-address to private: missing PNA header.
FAIL treat-as-public-address to private: success. assert_equals: expected "loaded" but got "timeout"
PASS treat-as-public-address to public: no preflight required.
PASS local to local, grandparent navigates: no preflight required.
FAIL public to local, grandparent navigates: failure. timeout adding grandchild
FAIL public to local, grandparent navigates: success. timeout adding grandchild
